assignment_1/requirements.txt
python-dotenv
slackclient
flask
slackeventsapi
pyenchant
truecase
geonamescache
spacy